requirements.txt
pytest==5.4.2
numpy~=1.18.2
neo4j~=1.7.6
pyyaml~=5.3.1
